reduce to one script.js file
